requirements_deploy.txt
# Core Dependencies for Deployment
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# AI/ML Dependencies
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.196
opencv-python==4.8.1.78
numpy==1.26.4
Pillow>=9.0.0

# Web Framework
streamlit==1.28.1
streamlit-option-menu==0.3.6
streamlit-aggrid==0.3.4.post3

# Data Processing
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0

# Configuration and Utilities
pyyaml>=6.0
python-dotenv>=0.19.0
pydantic>=1.10.0
httpx>=0.24.0

# Image Processing
imageio>=2.19.0
scikit-image>=0.19.0

# Logging and Monitoring
loguru>=0.6.0
psutil>=5.9.0

# Production Dependencies
gunicorn>=20.1.0
